Namespace optimize
