+++ Nager.AmazonProductAdvertising.Monitor/AuthenticationDialog.cs
@@ -1,12 +1,5 @@
-﻿using Nager.AmazonProductAdvertising.Model;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
+﻿using System;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
 namespace Nager.AmazonProductAdvertising.Monitor
