+++ Nager.AmazonProductAdvertising.TestConsole/Program.cs
@@ -20,7 +20,7 @@ namespace Nager.AmazonProductAdvertising.TestConsole
             authentication.AccessKey = accessKey;
             authentication.SecretKey = secretKey;
 
-            //ItemLookupRequest(authentication);
+            //LookupCheck.ItemLookupRequest(authentication);
             //ItemSearchRequest(authentication);
             //CustomItemSearchRequest1(authentication);
             //CustomItemSearchRequest2(authentication);
@@ -29,28 +29,6 @@ namespace Nager.AmazonProductAdvertising.TestConsole
             CreateCart1(authentication);
 
             Console.ReadLine();
-        }
-
-        static void ItemLookupRequest(AmazonAuthentication authentication)
-        {
-            Console.WriteLine("ItemLookup");
-            Console.WriteLine("------------------------------------------");
-
-            var wrapper = new AmazonWrapper(authentication, AmazonEndpoint.DE);
-            var result = wrapper.Lookup("B007KKKJYK");
-
-            Console.WriteLine("{0}", result.Items.Item[0].ItemAttributes.Title);
-
-            Console.WriteLine("Multi Lookup");
-            Console.WriteLine("------------------------------------------");
-
-            var result2 = wrapper.Lookup(new string[] { "B007KKKJYK", "B004MKNBJG" });
-            foreach (var item in result2.Items.Item)
-            {
-                Console.WriteLine(item.ItemAttributes.Title);
-            }
-
-            Console.WriteLine("------------------------------------------");
         }
 
         static void ItemSearchRequest(AmazonAuthentication authentication)
